backend/requirements.txt
django
djangorestframework
django-cors-headers
pymysql
djangorestframework-jwt